Final Part 3

+++ src/main/java/ge/epam/staffmanagement/controller/DepartmentController.java
@@ -25,7 +25,7 @@ public class DepartmentController {
     }
 
     @GetMapping("/id")
-    public Department getDepartmentById(int id) {
+    public Department getDepartmentById(long id) {
         return departmentService.getDepartmentById(id);
     }
 }
